MAKE LISTS INTO HORIZONTAL NAV BARS ON BOTH SIDES
I made the list items into inline-block elements to create
horizontal navigation menus. Then in order to push both
nav menus to either side of the page, I chose to use float
as a way to experiement with CSS float. However, notice
that the Google Image is now pushing all the way to the top
of the document because float elements are not considered in
the calculation of height in the parent container. Fix this
problem next either with clear fix, or overflow: hidden.